requirements.txt
streamlit==1.34.0
groq==0.8.0
sentence-transformers==2.5.0
huggingface-hub==0.16.4
faiss-cpu==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0
numpy==1.24.3
pandas==2.1.1
